Checkpointing profiling handler work.
The way the profilers work, they have to be invoked such that the frame
where profiling starts is not already running and such that the profiler
is the parent frame of that frame.  In order to get profiling of
existing/already-running frames, I might try to borrow a page from the
pure python profiler and fake the existing frames in the stack for the
profiler, but not today.